refactor(parser)!: relocate ParsedCommit object to better naming convention
the previous name of `token.py` was a copy/paste error and should of had a name relevant
to the classes inside of the module. This change moves the file and updates all internal
imports.

BREAKING CHANGE: The `ParsedCommit` class and `ParseError` class have moved to a more
relevant file name which changes the import path for those whom have written a custom
parser to use with Python Semantic Release.  To migrate, update the import path of
`ParsedCommit` and `ParseError` to use the `semantic_release.commit_parser.util` module.