fix(parsed-commit)!: fix commit objects immutability
BREAKING CHANGE: The `ParsedCommit` object data types have changed for the descriptions
and breaking descriptions attributes. They are now tuples of strings which makes the
parsed commit object fully immutable which was the original intent. This effects custom
parsers directly as the constructor will now fail if you pass the previous data type
`list[str]` as the parameters. Recommend wrapping your list in the `tuple()` constructor
during the `ParsedCommit()` constructor.

BREAKING CHANGE: The `ParsedCommit` object's attributes have changed order so in the case
your custom changelog is not passing the parameters as keyword arguments then likely the
initialization of the object will fail. The order changed because we added new defaults
to the breaking descriptions tuple and forced the commit attribute to be defined earlier
in the args list.