refactor(release-history)!: change data structure of commit history object
BREAKING CHANGE: This release swaps the long name of the commit type to the formal commit
type name as the key value into the `released[Version(1.0.0)]["elements"]` dictionary. Many
changelogs currently use a tuple unpack to grab the `type_` and then directly use it for the
subheading. Although type is still there it will be the shorthand value that is in the commit
message traditionally. To maintain the same functionality to populate headings with the longer
human-friendly names, use the new `commit.category` field instead. This change was made to
accommodate custom commit types introduced by user configuration which do not have their own
internal heading provided by PSR. Instead, custom changelogs can now just do a look up of the
formal type in a jinja dictionary to handle the human-friendly translation or use the category
field when you have a custom parser.