fix(parser-conventional)!: change parser option `allowed_tags` to `other_allowed_tags`
BREAKING CHANGE: The `allowed_tags` option was a bit misleading to new adopters as it
required both a new tag added to either `minor_tags` or `patch_tags` also to be included
in the `allowed_tags` list. Originally, `allowed_tags` was the list of where users could
define the additional tags that were still valid but are not configured to cause a version
bump. It had the side effect of requiring the inclusion of all tags even ones already
configured. It is easy in code to link the sets together, but not easy for users who are
defining these values within a static configuration file. Therefore, the old `allowed_tags`
was changed to `other_allowed_tags` with the change in the name denoting that it only needs
to have tags that are not otherwise defined in `minor_tags` or `patch_tags`. The previous
`allowed_tags` has now become a read-only property value that returns a wholistic list of
all tags configured in order from most to least significant.