fix(github-action)!: resolve command injection vulnerability in action script
Prevents the malicious execution of arbitrary code when a command injection is
defined in the action yaml as part of the action parameter specification. Low
impact exploitation if proper least privilege is implemented and pull request
code is reviewed before merging to a release branch.

BREAKING CHANGE: The `root_options` action input parameter has been removed
because it created a command injection vulernability for arbitrary code to
execute within the container context of the GitHub action if a command injection
code was provided as part of the `root_options` parameter string. To eliminate
the vulnerability, each relevant option that can be provided to `semantic-release`
has been individually added as its own parameter and will be processed individually
to prevent command injection. Please review our `Github Actions Configuration`__
page on the Python Semantic Release Documentation website to review the newly available
configuration options that replace the `root_options` parameter.

__ https://python-semantic-release.readthedocs.io/en/stable/automatic-releases/github-actions.html